pyproject.toml
[project]
name = "secbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=1.1.404",
    "langchain>=0.3.0",
    "openai>=1.45.0",
    "pre-commit>=3.8.0",
    "pyproject-toml>=0.0.10",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.5",
]

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
